changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11800 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Control/Ramper/ORRampItem.h
@@ -97,7 +97,7 @@
 - (NSMutableArray*) wayPoints;
 - (void) setWayPoints:(NSMutableArray*)someWayPoints;
 - (ORWayPoint*) wayPoint:(int)index;
-- (unsigned) wayPointCount;
+- (NSUInteger) wayPointCount;
 - (void) removeWayPoint:(id)aWayPoint;
 - (ORWayPoint*) currentWayPoint;
 - (BOOL) isRunning;
